Refactor for Directus 10
