--- .env.example
@@ -1,1 +1,0 @@
-DIRECTUS_EXTENSIONS_PATH=/directus/extensions/
