+++ .gitignore
@@ -1,4 +1,4 @@
-.DS_Store
-.env
 node_modules/
 dist/
+yarn.lock
+.DS_Store
