--- directus-module-publish-netlify/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@dwaring87/directus-module-publish-node",
-  "version": "1.0.0",
-  "description": "A Directus v9 module for building a local site from the Directus app",
-  "main": "src/index.js",
-  "scripts": {
-    "build": "npx directus-extension build",
-    "deploy": "DOTENV_CONFIG_PATH=../.env node scripts/deploy.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "github.com:dwaring87/directus-extension-publish-node"
-  },
-  "author": "David Waring <dev@davidwaring.net>",
-  "license": "MIT",
-  "devDependencies": {
-    "@directus/extensions-sdk": "^9.0.0-rc.98"
-  },
-  "directus:extension": {
-    "type": "module",
-    "path": "dist/index.js",
-    "source": "src/index.js",
-    "host": "^9.0.0-rc.98",
-    "hidden": false
-  }
-}
