+++ package.json
@@ -1,14 +1,45 @@
 {
-    "name": "@dwaring87/directus-extension-publish-netlify",
-    "workspaces": [
-        "./directus-module-publish-netlify",
-        "./directus-endpoint-publish-netlify"
-    ],
-    "scripts": {
-        "build": "npm run build --workspaces",
-        "deploy": "npm run deploy --workspaces"
+  "name": "directus-extension-publish-netlify",
+  "description": "A Directus extension that can be used to manage and run Netlify builds of your static site after updating your content in Directus",
+  "license": "MIT",
+  "icon": "extension",
+  "version": "1.0.0",
+  "keywords": [
+    "directus",
+    "directus-extension",
+    "directus-custom-bundle",
+    "netlify",
+    "static-site-generator"
+  ],
+  "type": "module",
+  "directus:extension": {
+    "type": "bundle",
+    "path": {
+      "app": "dist/app.js",
+      "api": "dist/api.js"
     },
-    "dependencies": {
-        "dotenv": "^10.0.0"
-    }
+    "entries": [
+      {
+        "type": "endpoint",
+        "name": "directus-extension-publish-netlify-endpoint",
+        "source": "src/endpoint/index.js"
+      },
+      {
+        "type": "module",
+        "name": "directus-extension-publish-netlify-module",
+        "source": "src/module/index.js"
+      }
+    ],
+    "host": "^10.3.0"
+  },
+  "scripts": {
+    "build": "directus-extension build",
+    "dev": "directus-extension build -w --no-minify",
+    "link": "directus-extension link",
+    "add": "directus-extension add"
+  },
+  "devDependencies": {
+    "@directus/extensions-sdk": "10.3.0",
+    "vue": "^3.4.14"
+  }
 }
