+++ src/module/module.vue
@@ -1,0 +1,10 @@
+<template>
+	<private-view title="My Custom Module">
+		<h2>Title</h2>
+		<p>Content goes here!</p>
+	</private-view>
+</template>
+
+<script>
+export default {};
+</script>
